Exploratory data analysis-Retail
n this task, we will be performing ‘Exploratory Data Analysis’ on the provided dataset ‘SampleSuperstore’ and draw insights to understand the sales performance of the store. We will analyze the data to identify the weak areas where the store can work to make more profit. This task will involve data cleaning, data visualization, and statistical analysis.